took out initialization of Dotenv in config/application.rb
had to take out manual initialization of Dotenv in this file b/c the Dotenv
gem is only installed in development & test so Dotenv can't be initialized
in production (on Heroku), leading to a there a compilation error when running rake.
Also, Dotenv runs automatically during 'before_configuration' callback during
the initialization of Applicaion defined in config/application.rb and there
isn't a need to do initialize Dotenv beforehand.
https://www.rubydoc.info/gems/dotenv-rails/2.1.1